requirements.txt
ipython
numpy
openai
opencv_python
Pillow
streamlit
